added Step3 template to Authorization
