+++ src/Navigation/styled.js
@@ -2,29 +2,34 @@ import React from 'react-router-dom'
 import styled from "styled-components";
 
 export const NavBar = styled.div`
-background-color:white
-padding:1%;
-`;
+  display: flex;
+  background-color:white;
+  padding:1%;
+`
 
 export const NavLinks = styled.div`
-width:15%;
-display:flex;
-justify-content: space-between;
-`;
+  width:15%;
+  display:flex;
+  justify-content: space-evenly;
+  align-items; center
+`
 
 export const NavBtn = styled.button`{
- width: @width;
- height: @height;
- color: #2B7BE3;
- font-weight: 500;
- text-align: center;
- background-color: #F3FSF9;
- border: 1px solid #2B7BE3;
- border-radius: 5px;
- cursor: pointer;
- font-family: ‘Rubik’, sans-serif;
- &:hover {
-   background-color: F3FSF9;
-   color: #E67271
-   border: none;
- }`
+  width: @width;
+  height: @height;
+  color: #2B7BE3;
+  font-weight: 500;
+  text-align: center;
+  background-color: #F3FSF9;
+  border: 1px solid #2B7BE3;
+  border-radius: 5px;
+  cursor: pointer;
+  font-family: ‘Rubik’, sans-serif;
+  &:hover {
+    background-color: F3FSF9;
+    color: #E67271
+    border: none;
+  }
+  marign: 10px;
+  
+`
